ENG-1360: CassandraKeyValue app doesn't stop after reading/writing requested number of keys.
Summary:
Fixed key counters and metrics tracker handling.
More details in https://yugabyte.atlassian.net/browse/ENG-1360.

Test Plan: Run CassandraKeyValue app, verify that writer/reader threads are stopping after generating requested load.

Reviewers: karthik, mikhail

Reviewed By: mikhail

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2062

+++ java/yb-loadtester/src/main/java/com/yugabyte/sample/common/IOPSThread.java
@@ -48,7 +48,7 @@ public class IOPSThread extends Thread {
     return ioThreadFailed;
   }
 
-  public int numOps() {
+  public long numOps() {
     return app.numOps();
   }
